Bug 1094272 - Use a single process. r=jgriffin
The child process has sometimes thrown exceptions and crashed,
but the main process continues to consume messages and write
them to the shared queue, thus losing them.  There's no real
need to have multiple processes; instead, we should use the
fact that messages will be queued in RabbitMQ if
pulsetranslator gets stuck on one for a while.

There's a possibility we'll get stuck too often and the
queue will grow so large as to be deleted by PulseGuardian,
but we'll see.

In addition,
* fix an exception when platform field is None
* add usage instructions to README
* clean up miscellaneous bits
* recreate publisher on each iteration.  It's not strictly
  necessary, but it's cleaner to create a new one each time
  through the publish retry loop.
* stop printing tracebacks.