Bug 1094272 - Use a single process.
The child process has sometimes thrown exceptions and crashed,
but the main process continues to consume messages and write
them to the shared queue, thus losing them.  There's no real
need to have multiple processes; instead, we should use the
fact that messages will be queued in RabbitMQ if
pulsetranslator gets stuck on one for a while.

In addition,
* fix an exception when platform field is None
* add usage instructions to README
* clean up miscellaneous bits